Pull out TestHandler
